feat: Fix min Table ux issues. (#4125)
